Update vpd_runtime_test.cpp
Modification supporting update to temperature conversion classes.

+++ Transpiration/vpd_runtime_test.cpp
@@ -1,6 +1,6 @@
 //
 // vpd_runtime_test.cpp
-// Version timestamp: 9-4-2018, 4:23 PM
+// Version timestamp: 2-10-2020, 8:26 AM
 //
 // Attribution : Copyright (c) 2018 Northern_Loki (sha256::6F290BF833967127BE26C92C8F6B1C1A3949C55A7EABCEF3ECC785CD2D38D30D)
 // License is granted under the Creative Commons Attribution-ShareAlike 4.0 International.  https://creativecommons.org/licenses/by-sa/4.0/
@@ -130,7 +130,7 @@ float vpd_runtime_test::check_vpd(celcius temperature, celcius leaf_temperature,
 float vpd_runtime_test::check_vpd_and_print(vpd vapor_pressure, float min_vpd, float max_vpd)
 {
 	double vapor_pressure_deficit = vapor_pressure.vapor_pressure_deficit;
-	print_table(vapor_pressure.degrees_celcius, vapor_pressure.humidity*100, (vapor_pressure.degrees_celcius_leaf - vapor_pressure.degrees_celcius), min_vpd, max_vpd);
+	print_table(vapor_pressure.degrees_celcius, vapor_pressure.humidity*100, (float) (vapor_pressure.degrees_celcius_leaf - vapor_pressure.degrees_celcius), min_vpd, max_vpd);
 	return (print_status(vapor_pressure_deficit, min_vpd, max_vpd));
 }
 
@@ -145,7 +145,7 @@ float vpd_runtime_test::check_vpd_and_print(celcius temperature, float humidity,
 float vpd_runtime_test::check_vpd_and_print(celcius temperature, celcius leaf_temperature, float humidity, float min_vpd, float max_vpd)
 {
 	double vapor_pressure_deficit = calc_boundary(temperature, leaf_temperature, humidity);
-	print_table(temperature, humidity, (leaf_temperature - temperature), min_vpd, max_vpd);
+	print_table(temperature, humidity, (float)(leaf_temperature - temperature), min_vpd, max_vpd);
 	float result = print_status(vapor_pressure_deficit, min_vpd, max_vpd);
 	return (result);
 }
